pyproject.toml
[project]
name = "nonebot-plugin-arcaea-sticker"
version = "0.3.3"
description = "Arcaea sticker generator plugin for NoneBot2"
authors = [
    { name = "JQ-28", email = "480352716@qq.com" }
]
dependencies = [
    "nonebot2>=2.0.0",
    "nonebot-adapter-onebot>=2.0.0",
    "nonebot-plugin-htmlrender>=0.2.0",
    "nonebot-plugin-localstore>=0.7.0",
    "Pillow>=9.0.0",
    "jinja2>=3.0.0",
    "anyio>=3.0.0",
    "yarl>=1.0.0",
    "pydantic>=2.0.0",
    "playwright>=1.20.0"
]
requires-python = ">=3.8,<4.0"
readme = "README.md"
license = { text = "MIT" }

keywords = ["nonebot2", "plugin", "arcaea", "sticker"]
packages = [
    { include = "nonebot_plugin_arcaea_sticker" }
]

[project.urls]
homepage = "https://github.com/JQ-28/nonebot-plugin-arcaea-sticker"
repository = "https://github.com/JQ-28/nonebot-plugin-arcaea-sticker"
documentation = "https://github.com/JQ-28/nonebot-plugin-arcaea-sticker#readme"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["nonebot_plugin_arcaea_sticker"]

[tool.pdm]
[tool.pdm.build]
includes = ["nonebot_plugin_arcaea_sticker"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"